WIP [OPP-1384] flytte retry ut til egen util
